refactor(frontend): tidy audit event table, remove drop-down menu and add toggle for all events
